// src/types/types.ts
export type WeatherInfo = {
  maxTemp: number;
  minTemp: string;
  date: string;
  iconCode: string;
  isDay: boolean;
};